+++ payments.php
@@ -41,7 +41,7 @@ while($record=mysqli_fetch_array($run_cart)){
 	
 	$pro_qty = $record['qty'];
 		
-	$sub_total = (int)$record['p_price']*(int)$pro_qty;
+	$sub_total = (float)$record['p_price']*(int)$pro_qty;
 		
 	$total += $sub_total;
